feat: global/local pos

+++ src/main/kotlin/com/timur/nimcube/command/CuboidCommand.kt
@@ -6,6 +6,7 @@ import org.bukkit.command.Command
 import org.bukkit.command.CommandExecutor
 import org.bukkit.command.CommandSender
 import org.bukkit.entity.Player
+import org.joml.Vector3d
 import org.joml.Vector3f
 import java.lang.foreign.Arena
 
@@ -26,7 +27,7 @@ class CuboidCommand(val plugin: Nimcube) : CommandExecutor {
 
         Arena.ofConfined().use { arena ->
             val worldIndex = nim.createWorld()
-            val cuboid = nim.createCuboid(arena, worldIndex, Vector3f(10f, 20f, 30f))
+            val cuboid = nim.createCuboid(arena, worldIndex, Vector3d(10.0, 20.0, 30.0))
             val pos = nim.getCuboidPos(arena, worldIndex, cuboid)
             println("world: $worldIndex")
             println("cuboid: $cuboid")
